Finish first draft of infographic

+++ Makefile
@@ -1,5 +1,11 @@
 CID = 01725740
 
+fonts:
+	mkdir fonts/
+	curl -X GET --user-agent "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_7) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.114 Safari/537.36" "https://fonts.google.com/download?family=Lato" --output fonts/Lato.zip
+	unzip fonts/Lato.zip -d fonts/
+	rm fonts/Lato.zip
+
 clean:
 	echo "TODO"
 	cd reports/
@@ -9,10 +15,15 @@ clean:
 package:
 	git ls-files
 
-infographic:
+infographic: fonts
+	# make sure the csv is available, from which the plots are made
+	test -s data/gdpr_compliancy_data_full.csv || { echo "The GDPR compliancy data CSV was not found! Run the SQL query before attempting to build the infographic."; exit 1; }
+	# create the plots
+	Rscript analyses/plotting/gdpr_violations_by_country.R
+	Rscript analyses/plotting/websites_of_the_internet.R
 	# convert the plots
 	inkscape --export-pdf=outputs/internet_waffle.pdf outputs/internet_waffle.eps
 	inkscape --export-pdf=outputs/gdpr_by_country.pdf outputs/gdpr_by_country.eps
-	# compile and export
+	# compile the infographic and move a copy it to root (lualatex needed for the fontenc package)
 	latexmk -pdf -lualatex -output-directory="reports/build/" -verbose -synctex=1 -shell-escape reports/infographic.tex
 	cp reports/build/infographic.pdf $(CID)-submission.pdf
